% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_power.R
\name{prop_power}
\alias{prop_power}
\title{Power and sample size for 2 proportions}
\usage{
prop_power(n, n1, n2, p1, p2, fraction = 0.5, alpha = 0.05,
  power = NULL, alternative = c("two.sided", "one.sided"),
  strict = FALSE, tol = .Machine$double.eps^0.25, odds.ratio,
  percent.reduction)
}
\description{
Calculate power and sample size for comparison of 2 proportions for both balanced and ubalanced designs.
}
