% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tavolo.R
\name{tavolo.matrix}
\alias{tavolo.matrix}
\title{Create 2 x k frequency tables}
\usage{
\method{tavolo}{matrix}(df, dnn = NULL, rev = c("neither", "rows",
  "columns", "both"), ...)
}
\arguments{
\item{df}{a 2 x k frequency matrix.}

\item{dnn}{optional character vector of dimension names.}

\item{rev}{character string indicating whether to switch row or column order, possible options are "neither", "rows", "columns", or "both". The default is "neither".}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\item{tab}{2 x k frequency table}
}
\description{
Helper function for creating 2 x k frequency tables.
}
\examples{
tavolo.matrix(matrix(c(23, 45, 67, 12), nrow = 2, ncol = 2), rev = "both")
}
