% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tavolo.R
\name{tavolo}
\alias{tavolo}
\title{Create 2 x k frequency tables}
\usage{
tavolo(df, ...)
}
\arguments{
\item{df}{a dataframe with binary variable y and categorical variable x or a 2 x k frequency table/matrix. If a table or matrix, x and y must be NULL. Used to select method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\item{tab}{2 x k frequency table}
}
\description{
Helper function for creating 2 x k frequency tables.
}
\examples{
trial <- data.frame(disease = c(rep("yes", 2), rep("no", 2)),
                    treatment = c(rep(c("estrogen", "placebo"), 2)),
                    count = c(751, 623, 7755, 7479))

tavolo(trial, treatment, disease, count)

}
