% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_deviance.R
\name{gof_deviance}
\alias{gof_deviance}
\title{Conduct goodness of fit for logistic regression using deviance}
\usage{
gof_deviance(working, saturated)
}
\arguments{
\item{saturated}{}
}
\description{
Conduct goodness of fit for logistic regression using deviance
}
