% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matched_or.R
\name{matched_or}
\alias{matched_or}
\title{Matched pairs odds ratio and confidence interval}
\usage{
matched_or(df, ...)
}
\arguments{
\item{df}{a dataframe with binary variables x and y or a 2 x 2 frequency table/matrix. If a table or matrix, x and y must be NULL. Used to select method.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{binary vector, used as rows for frequency table and calculations; ignored if df is a table.}

\item{y}{binary vector, used as columns for frequency table and calculations; ignored if df is a table.}

\item{weight}{an optional vector of count weights. Must be blank if df is a table or matrix.}

\item{alpha}{level of significance for confidence interval.}

\item{rev}{reverse order of cells. Options are "row", "columns", "both", and "neither" (default).}
}
\value{
a list with class "matched_or" with the following components:

\item{tab}{2x2 table using for calculating risk difference}
\item{or}{dataframe with columns corresponding to matched-pairs OR, lower bound, and upper bound of CI}
\item{conf.level}{specified confidence level}
}
\description{
Matched pairs odds ratio and confidence interval
}
\examples{

gene <- data.frame(pair = seq(1:35),
                   ulcer = rbinom(35, 1, .7),
                   healthy = rbinom(35, 1, .4))

matched_or(gene, ulcer, healthy)

}
