% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_test.R
\name{prop_test}
\alias{prop_test}
\title{Tests for equality of proportions}
\usage{
prop_test(x, n, p = NULL, method = c("wald", "wilson", "agresti-couli",
  "jeffreys", "modified wilson", "wilsoncc", "modified jeffreys",
  "clopper-pearson", "arcsine", "logit", "witting", "pratt"),
  alternative = c("two.sided", "less", "greater"), conf.level = 0.95,
  correct = FALSE, exact = FALSE)
}
\description{
Run tests for equality of proportions
}
