% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_test.R
\name{prop_test}
\alias{prop_test}
\title{Tests for equality of proportions}
\usage{
prop_test(x, n, p = NULL, method = c("wald", "wilson", "agresti-couli",
  "jeffreys", "modified wilson", "wilsoncc", "modified jeffreys",
  "clopper-pearson", "arcsine", "logit", "witting", "pratt"),
  alternative = c("two.sided", "less", "greater"), conf.level = 0.95,
  correct = FALSE, exact = FALSE)
}
\arguments{
\item{x}{a vector of counts, a one-dimensional table with two entries, or a two-dimensional table with 2 columns.}

\item{n}{a vector of counts of trials, ignored if x is a matrix or table.}

\item{p}{a vector of probabilities, length must be same as the number of groups specified by x.}

\item{method}{a character string indicating method for calculating confidence interval, default is "wald".}

\item{alternative}{character string specifiying the alternative hypothesis. Possible options are "two.sided" (default),
"greater", or "less".}

\item{conf.level}{confidence level for confidence interval, default is 0.95.}

\item{correct}{a logical indicating whether Yate's continuity correction should be applied.}

\item{exact}{a logical indicating whether to output exact p-value, ignored if k-sample test.}
}
\value{
a list with class "prop_test" containing the following components:

\item{x}{number of successes}
\item{n}{number of trials}
\item{p}{null proportion}
\item{statistic}{the value of Pearson's chi-squared test statistic}
\item{p_value}{p-value corresponding to chi-squared test statistic}
\item{df}{degrees of freedom}
\item{method}{the method used to calculate the confidence interval}
\item{method_ci}{confidence interval calculated using specified method}
\item{exact_ci}{exact confidence interval}
\item{exact_p}{p-value from exact test}
}
\description{
Run tests for equality of proportions
}
\examples{
prop_test(7, 50, method = "wald", p = 0.2)
prop_test(7, 50, method = "wald", p = 0.2, exact = TRUE)

vietnam <- data.frame(
   service = c(rep("yes", 2), rep("no", 2)),
   sleep = c(rep(c("yes", "no"), 2)),
   count = c(173, 160, 599, 851)
)

sleep <- xtabs(count ~ service + sleep, data = vietnam)
prop_test(sleep)

}
