% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskdiff.R
\name{riskdiff}
\alias{riskdiff}
\title{Risk difference}
\usage{
riskdiff(x, y = NULL, weight = NULL, conf.level = 0.95,
  rev = c("neither", "rows", "columns", "both"), dnn = NULL)
}
\description{
Calculate risk difference and 95 percent confidence interval using Wald method
}
