% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskdiff.R
\name{riskdiff}
\alias{riskdiff}
\title{Risk difference}
\usage{
riskdiff(df = NULL, x = NULL, y = NULL, weight = NULL,
  conf.level = 0.95, rev = c("neither", "rows", "columns", "both"),
  dnn = NULL)
}
\arguments{
\item{df}{a one-dimensional dataframe containing counts or a two-dimensional dataframe/matrix of counts.
If a matrix, x and y must be NULL}

\item{x}{a vector of counts or a 2x2 matrix of counts. If a matrix, df and y must be NULL}

\item{y}{an optional vector of counts}

\item{weight}{an optional vector of count weights}

\item{conf.level}{confidence level for confidence interval, default is 0.95}

\item{rev}{reverse order cells. Options are "row", "columns", "both", and "neither" (default)}

\item{dnn}{optional character vector of dimension names}
}
\value{
a list with class "rdiff" containing the following components:

\item{rd}{risk difference}
\item{conf.level}{specified confidence level}
\item{ci}{calculated confidence interval}
\item{p1}{proportion one}
\item{p2}{proportion two}
\item{tab}{2x2 table using for calculating risk difference}
}
\description{
Calculate risk difference and 95 percent confidence interval using Wald method.
}
\examples{
trial <- data.frame(
  disease = c(rep("yes", 2), rep("no", 2)),
  treatment = c(rep(c("estrogen", "placebo"), 2)),
  count = c(751, 623, 7755, 7479))

riskdiff(trial, "treatment", "disease", weight = "count", rev = "columns")

exposed <- round(runif(1000))
diseased <- round(runif(1000))

riskdiff(x = exposed, y = diseased, rev = "both")

}
