% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tobyto.R
\name{tobyto}
\alias{tobyto}
\alias{tobyto.matrix}
\alias{tobyto.table}
\alias{tobyto.data.frame}
\title{Create frequency table}
\usage{
tobyto(...)

\method{tobyto}{matrix}(..., rev = c("neither", "rows", "columns",
  "both"))

\method{tobyto}{table}(..., rev = c("neither", "rows", "columns",
  "both"))

\method{tobyto}{data.frame}(..., rev = c("neither", "rows", "columns",
  "both"))
}
\arguments{
\item{...}{argument calls from \code{riskdiff()}}

\item{rev}{character string indicating whether to switch row or column order, possible options are "neither", "rows", "columns", or "both". The default is "neither".}
}
\description{
Create frequency table
}
